Move vp check function
